+++ Example/WICompressExample/WICompressExample/WICompressExampleViewModel.swift
@@ -75,9 +75,9 @@ final class WICompressExampleViewModel {
         logger.info("Original format: \(imageGroup.format)")
         
         do {
-            let compressedData = try WICompress.compress(
+            let compressedData = try WICompress.process(
                 imageGroup.rawData,
-                options: WICompressOptions(quality: .compression(0.7))
+                using: WIImageProcess(quality: 0.7)
             )
             guard let compressedUIImage = UIImage(data: compressedData) else {
                 logger.error("Compression output could not be decoded!")